+++ authserver/src/internal/server/handler_admin_client_oauth2_flows.go
@@ -28,7 +28,7 @@ func (s *Server) handleAdminClientOAuth2Get() http.HandlerFunc {
 			s.internalServerError(w, r, err)
 			return
 		}
-		client, err := s.databasev2.GetClientById(nil, int64(id))
+		client, err := s.databasev2.GetClientById(nil, id)
 		if err != nil {
 			s.internalServerError(w, r, err)
 			return
@@ -99,7 +99,7 @@ func (s *Server) handleAdminClientOAuth2Post() http.HandlerFunc {
 			return
 		}
 
-		client, err := s.databasev2.GetClientById(nil, int64(id))
+		client, err := s.databasev2.GetClientById(nil, id)
 		if err != nil {
 			s.internalServerError(w, r, err)
 			return
