+++ authserver/src/internal/server/handler_admin_group_attributes_add.go
@@ -28,7 +28,7 @@ func (s *Server) handleAdminGroupAttributesAddGet() http.HandlerFunc {
 			s.internalServerError(w, r, err)
 			return
 		}
-		group, err := s.databasev2.GetGroupById(nil, int64(id))
+		group, err := s.databasev2.GetGroupById(nil, id)
 		if err != nil {
 			s.internalServerError(w, r, err)
 			return
